requirements.txt
Flask==2.1.2
GitPython==3.1.27
luigi==3.6.0
matplotlib==3.5.2
numpy==1.21.6
polars==1.16.0
pytest==8.3.5
python-dotenv==1.0.1
pytz==2021.3
requests==2.28.1
supabase==2.10.0
Werkzeug==2.1.2
